+++ src/components/backButton/backButton.css
@@ -1,0 +1,12 @@
+.back {
+  position: absolute;
+  top: 3%;
+  left: 2%;
+}
+
+.back button {
+  border: none;
+  background-color: transparent;
+  color: white;
+  font-size: 50px;
+}
