+++ src/components/home/home.css
@@ -3,14 +3,11 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-}
-
-.store-name {
-  color: white;
+  height: 100%;
 }
 
 .home-img {
-  margin: 2rem 2rem;
+  margin: 0 2rem;
   width: 15rem;
   border-radius: 20px;
   box-shadow:0px 0px 43px 0px rgba(196,46,255,0.9);
@@ -18,6 +15,13 @@
 
 .home-icon {
   color: rgb(203, 255, 13);
+}
+
+.presentation {
+  text-align: center;
+  width: 80%;
+  color: #fff;
+  margin: 1rem 0;
 }
 
 .categories {
@@ -71,4 +75,11 @@
 
 .categories a:nth-child(4) {
   background-color: rgb(55, 70, 186);
+}
+
+.footer {
+  position: absolute;
+  bottom: 0;
+  color: rgb(184, 182, 182);
+  font-size: 1.2rem;
 }
