+++ src/components/home/home.js
@@ -13,29 +13,34 @@ const HomeList = () => {
   }, []);
 
   return (
-    <div className="articles_container">
-
-      <h2 className="store-name">
-        Premium store
-        <i className="fa-solid fa-store home-icon" />
-      </h2>
-      <img className="home-img" src={home} alt="home" />
-      <div className="categories">
-        {categories && categories.map(
-          (category) => (
-            <Link to={`/category/${category.name}`} key={category.name} className={category.class}>
-              <p>{category.name}</p>
-              <span>
-                (
-                {category.productsCount}
-                )
-              </span>
-            </Link>
-          ),
-        )}
+    <>
+      <div className="articles_container">
+        <img className="home-img" src={home} alt="home" />
+        <p className="presentation">
+          <h2>We are open!</h2>
+          We are an online store with the newest technologie
+          and electronics
+        </p>
+        <div className="categories">
+          {categories && categories.map(
+            (category) => (
+              <Link to={`/category/${category.name}`} key={category.name} className={category.class}>
+                <p>{category.name}</p>
+                <span>
+                  (
+                  {category.productsCount}
+                  )
+                </span>
+              </Link>
+            ),
+          )}
+        </div>
+        <footer className="footer">
+          <span>Project build by Arturo</span>
+        </footer>
       </div>
-    </div>
 
+    </>
   );
 };
